Port sorting of floats to `core`
Before `num_traits::Float` was used, which does not support `no_std`
because it uses a lot of math functions (`cos`, `sin`, etc.) that are
not implemented in `core`. Those are however not needed for sorting
floats, so we implement a minimal trait required for sorting floats
inside this crate instead.

The `float` feature was left in for backwards compatibility. It does not
do anything, because now we can implement sorting of floats by default.